+++ front/package.json
@@ -67,6 +67,7 @@
     "@storybook/react": "^6.2.9",
     "@types/react": "^17.0.3",
     "@types/react-dom": "^17.0.3",
+    "@types/react-redux": "^7.1.16",
     "source-map-loader": "^2.0.1",
     "ts-loader": "^8.2.0",
     "typescript": "^4.2.4"
